Intent Parser Tests
This adds three intent parser tests for comparison, along with
a set intents that are duplicated as far as possible between
all three systems, and a set of requests to parse.

This demonstrates how an intent can be translated between the
these systems and also allows direct comparison of performance
between the three.