Added support for unspecified keywords
This allows the plugin developer to specify a keyword to be
matched in a template without specifying a list of matches
for that keyword. It also allows the user to use a word that
the developer did not include in a list, even if the
developer did specify a list of words associated with the
keyword.